refactor: split unit tests into focused files by functionality
Replaces 3 monolithic test files (render.test.ts, hooks.test.ts,
context-and-reconciler.test.ts) with 13 focused files, one per
feature area, while keeping all 132 tests intact.

New structure:
- render-html.test.ts       — HTML elements, buildNode, HTML defaults
- render-components.test.ts — plain function & generator components
- shown-prop.test.ts        — $shown conditional rendering
- use-state.test.ts         — useState hook
- use-ref.test.ts           — useRef hook
- use-id.test.ts            — useId hook
- use-memo.test.ts          — useMemo hook
- use-effect.test.ts        — useEffect hook
- use-resolve.test.ts       — useResolve + useResolveRaw
- use-render.test.ts        — useRender (both variants) + useResume
- ui-patch.test.ts          — startUIPatch/commitUIPatch, useUIPatch,
                              live-only reconcile (global + local)
- context.test.ts           — createContext/useContext
- reconciler.test.ts        — prop memoization, component renders component
- jsx.test.ts               — createElement (unchanged)

Closes #46
