--- src/__tests__/render.test.ts
@@ -1,1161 +1,0 @@
-import { createElement, Fragment } from '../jsx';
-import { render, buildNode } from '../render';
-import { useState, useEffect, useResolve, useResolveRaw, useMemo } from '../hooks';
-
-// jsdom is provided by jest-environment-jsdom (see jest.config.js)
-
-describe('render – HTML elements', () => {
-  let container: HTMLElement;
-
-  beforeEach(() => {
-    container = document.createElement('div');
-    document.body.appendChild(container);
-  });
-
-  afterEach(() => {
-    document.body.removeChild(container);
-  });
-
-  it('renders a plain element', () => {
-    render(createElement('div', null), container);
-    expect(container.firstChild).toBeInstanceOf(HTMLDivElement);
-  });
-
-  it('renders a text child', () => {
-    render(createElement('p', null, 'Hello World'), container);
-    expect(container.querySelector('p')!.textContent).toBe('Hello World');
-  });
-
-  it('renders nested elements', () => {
-    render(createElement('div', null, createElement('span', null, 'inner')), container);
-    expect(container.querySelector('span')!.textContent).toBe('inner');
-  });
-
-  it('applies className', () => {
-    render(createElement('div', { className: 'foo bar' }), container);
-    expect((container.firstChild as HTMLElement).className).toBe('foo bar');
-  });
-
-  it('applies arbitrary attributes', () => {
-    render(createElement('input', { type: 'text', placeholder: 'name' }), container);
-    const input = container.querySelector('input')!;
-    expect(input.getAttribute('type')).toBe('text');
-    expect(input.getAttribute('placeholder')).toBe('name');
-  });
-
-  it('applies event listeners and wraps in SyntheticEvent', () => {
-    const onClick = jest.fn();
-    render(createElement('button', { onClick }, 'click me'), container);
-    container.querySelector('button')!.click();
-    expect(onClick).toHaveBeenCalledTimes(1);
-    // The handler receives a SyntheticEvent, not the raw native event
-    const syntheticEvent = onClick.mock.calls[0][0];
-    expect(syntheticEvent).toHaveProperty('nativeEvent');
-    expect(syntheticEvent).toHaveProperty('type', 'click');
-    expect(typeof syntheticEvent.preventDefault).toBe('function');
-    expect(typeof syntheticEvent.stopPropagation).toBe('function');
-  });
-
-  it('applies inline styles', () => {
-    render(createElement('div', { style: { color: 'red', fontSize: '14px' } }), container);
-    const el = container.firstChild as HTMLElement;
-    expect(el.style.color).toBe('red');
-    expect(el.style.fontSize).toBe('14px');
-  });
-
-  it('updates changed style properties on rerender', () => {
-    let setStyle: ((s: Record<string, string>) => void) | null = null;
-
-    function* Styled() {
-      const [style, ss] = yield* useState<Record<string, string>>({ color: 'red' });
-      setStyle = ss;
-      return createElement('div', { style });
-    }
-
-    render(createElement(Styled as never, {}), container);
-    const el = container.querySelector('div') as HTMLElement;
-    expect(el.style.color).toBe('red');
-
-    setStyle!({ color: 'blue' });
-    expect(el.style.color).toBe('blue');
-  });
-
-  it('removes style properties that are no longer present on rerender', () => {
-    let setStyle: ((s: Record<string, string>) => void) | null = null;
-
-    function* Styled() {
-      const [style, ss] = yield* useState<Record<string, string>>({
-        color: 'red',
-        fontSize: '14px',
-      });
-      setStyle = ss;
-      return createElement('div', { style });
-    }
-
-    render(createElement(Styled as never, {}), container);
-    const el = container.querySelector('div') as HTMLElement;
-    expect(el.style.color).toBe('red');
-    expect(el.style.fontSize).toBe('14px');
-
-    setStyle!({ color: 'blue' });
-    expect(el.style.color).toBe('blue');
-    expect(el.style.fontSize).toBe('');
-  });
-
-  it('clears all styles when style prop is removed', () => {
-    let setProps: ((p: Record<string, unknown>) => void) | null = null;
-
-    function* Styled() {
-      const [props, sp] = yield* useState<Record<string, unknown>>({
-        style: { color: 'red', fontWeight: 'bold' },
-      });
-      setProps = sp;
-      return createElement('div', props);
-    }
-
-    render(createElement(Styled as never, {}), container);
-    const el = container.querySelector('div') as HTMLElement;
-    expect(el.style.color).toBe('red');
-    expect(el.style.fontWeight).toBe('bold');
-
-    setProps!({});
-    expect(el.style.color).toBe('');
-    expect(el.style.fontWeight).toBe('');
-  });
-
-  it('renders a Fragment with multiple children', () => {
-    render(
-      createElement(
-        'ul',
-        null,
-        createElement(
-          Fragment,
-          null,
-          createElement('li', null, 'one'),
-          createElement('li', null, 'two'),
-        ),
-      ),
-      container,
-    );
-    const items = container.querySelectorAll('li');
-    expect(items).toHaveLength(2);
-    expect(items[0].textContent).toBe('one');
-    expect(items[1].textContent).toBe('two');
-  });
-
-  it('skips null and undefined children', () => {
-    render(createElement('div', null, null, undefined, 'visible'), container);
-    expect(container.querySelector('div')!.textContent).toBe('visible');
-  });
-  it('sets input value as DOM property (not just attribute)', () => {
-    let setValue: ((v: string) => void) | null = null;
-
-    function* Controlled() {
-      const [val, sv] = yield* useState('initial');
-      setValue = sv;
-      return createElement('input', { type: 'text', value: val });
-    }
-
-    render(createElement(Controlled as never, {}), container);
-    const input = container.querySelector('input') as HTMLInputElement;
-    expect(input.value).toBe('initial');
-
-    setValue!('updated');
-    expect(input.value).toBe('updated');
-
-    setValue!('');
-    expect(input.value).toBe('');
-  });
-
-  it('sets checkbox checked as DOM property', () => {
-    let setChecked: ((v: boolean) => void) | null = null;
-
-    function* CheckBox() {
-      const [checked, sc] = yield* useState(false);
-      setChecked = sc;
-      return createElement('input', { type: 'checkbox', checked });
-    }
-
-    render(createElement(CheckBox as never, {}), container);
-    const cb = container.querySelector('input') as HTMLInputElement;
-    expect(cb.checked).toBe(false);
-
-    setChecked!(true);
-    expect(cb.checked).toBe(true);
-
-    setChecked!(false);
-    expect(cb.checked).toBe(false);
-  });
-});
-
-describe('render – plain function components', () => {
-  let container: HTMLElement;
-
-  beforeEach(() => {
-    container = document.createElement('div');
-    document.body.appendChild(container);
-  });
-
-  afterEach(() => {
-    document.body.removeChild(container);
-  });
-
-  it('calls the function and renders the returned JSX', () => {
-    function Greeting({ name }: { name: string }) {
-      return createElement('h1', null, `Hello, ${name}!`);
-    }
-    render(createElement(Greeting as never, { name: 'World' }), container);
-    expect(container.querySelector('h1')!.textContent).toBe('Hello, World!');
-  });
-
-  it('renders an empty host when the component returns null', () => {
-    function Empty() {
-      return null;
-    }
-    render(createElement(Empty as never, {}), container);
-    const host = container.firstChild as HTMLElement;
-    expect(host.childElementCount).toBe(0);
-  });
-
-  it('renders an empty host when the component returns undefined', () => {
-    function Empty() {
-      return undefined;
-    }
-    render(createElement(Empty as never, {}), container);
-    const host = container.firstChild as HTMLElement;
-    expect(host.childElementCount).toBe(0);
-  });
-});
-
-describe('render – generator components', () => {
-  let container: HTMLElement;
-
-  beforeEach(() => {
-    container = document.createElement('div');
-    document.body.appendChild(container);
-  });
-
-  afterEach(() => {
-    document.body.removeChild(container);
-  });
-
-  it('renders a generator component that returns JSX', () => {
-    function* Greeting({ name }: { name: string }) {
-      return createElement('h2', null, `Hi, ${name}!`);
-    }
-    render(createElement(Greeting as never, { name: 'Alice' }), container);
-    expect(container.querySelector('h2')!.textContent).toBe('Hi, Alice!');
-  });
-
-  it('rerenders via useState setter', () => {
-    let setCount: ((v: number) => void) | null = null;
-
-    function* Counter() {
-      const [count, sc] = yield* useState(0);
-      setCount = sc;
-      return createElement('button', {}, String(count));
-    }
-
-    render(createElement(Counter as never, {}), container);
-    expect(container.querySelector('button')!.textContent).toBe('0');
-
-    setCount!(1);
-    expect(container.querySelector('button')!.textContent).toBe('1');
-
-    setCount!(5);
-    expect(container.querySelector('button')!.textContent).toBe('5');
-  });
-
-  it('rerenders when rerender() is called directly from an event handler', () => {
-    function* Counter(_props: Record<string, unknown>, rerender: () => void) {
-      const [count, setCount] = yield* useState(0);
-      return createElement(
-        'button',
-        {
-          onClick: () => {
-            setCount(count + 1);
-          },
-        },
-        String(count),
-      );
-    }
-
-    render(createElement(Counter as never, {}), container);
-
-    expect(container.querySelector('button')!.textContent).toBe('0');
-    container.querySelector('button')!.click();
-    expect(container.querySelector('button')!.textContent).toBe('1');
-    container.querySelector('button')!.click();
-    expect(container.querySelector('button')!.textContent).toBe('2');
-  });
-
-  it('renders generator components nested inside HTML elements', () => {
-    function* Label({ text }: { text: string }) {
-      return createElement('span', null, text);
-    }
-
-    render(
-      createElement(
-        'div',
-        { className: 'wrapper' },
-        createElement(Label as never, { text: 'nested' }),
-      ),
-      container,
-    );
-
-    expect(container.querySelector('span')!.textContent).toBe('nested');
-  });
-
-  it('passes children in props', () => {
-    function* Wrapper({ children }: { children: unknown }) {
-      return createElement('section', null, ...(children as never[]));
-    }
-
-    render(
-      createElement(Wrapper as never, {}, createElement('p', null, 'child content')),
-      container,
-    );
-
-    expect(container.querySelector('p')!.textContent).toBe('child content');
-  });
-});
-
-describe('render – generator components with useState', () => {
-  let container: HTMLElement;
-
-  beforeEach(() => {
-    container = document.createElement('div');
-    document.body.appendChild(container);
-  });
-
-  afterEach(() => {
-    document.body.removeChild(container);
-  });
-
-  it('useState persists value across re-renders', () => {
-    let setLabel: ((v: string) => void) | null = null;
-
-    function* Label() {
-      const [text, st] = yield* useState('initial');
-      setLabel = st;
-      return createElement('p', null, text);
-    }
-
-    render(createElement(Label as never, {}), container);
-    expect(container.querySelector('p')!.textContent).toBe('initial');
-
-    setLabel!('updated');
-    expect(container.querySelector('p')!.textContent).toBe('updated');
-
-    setLabel!('again');
-    expect(container.querySelector('p')!.textContent).toBe('again');
-  });
-
-  it('multiple useState calls maintain independent state', () => {
-    let setA: ((v: string) => void) | null = null;
-    let setB: ((v: number) => void) | null = null;
-
-    function* Multi() {
-      const [a, sa] = yield* useState('hello');
-      const [b, sb] = yield* useState(0);
-      setA = sa;
-      setB = sb;
-      return createElement('p', null, `${a}-${b}`);
-    }
-
-    render(createElement(Multi as never, {}), container);
-    expect(container.querySelector('p')!.textContent).toBe('hello-0');
-
-    setA!('world');
-    expect(container.querySelector('p')!.textContent).toBe('world-0');
-
-    setB!(42);
-    expect(container.querySelector('p')!.textContent).toBe('world-42');
-  });
-});
-
-describe('render – generator components with useResolve', () => {
-  let container: HTMLElement;
-
-  beforeEach(() => {
-    container = document.createElement('div');
-    document.body.appendChild(container);
-  });
-
-  afterEach(() => {
-    document.body.removeChild(container);
-  });
-
-  it('shows loading state while promise is pending', async () => {
-    let resolvePromise!: (data: string) => void;
-    const promise = new Promise<string>((res) => {
-      resolvePromise = res;
-    });
-
-    function* DataComp() {
-      const data = yield* useResolve(
-        {
-          fn: (_signal) => promise,
-          loading: createElement('span', { id: 'loading' }, 'Loading…'),
-          error: createElement('span', { id: 'error' }, 'Error'),
-        },
-        [],
-      );
-      return createElement('span', { id: 'data' }, data);
-    }
-
-    render(createElement(DataComp as never, {}), container);
-    expect(container.querySelector('#loading')).not.toBeNull();
-    expect(container.querySelector('#data')).toBeNull();
-
-    resolvePromise('Hello World');
-    await promise;
-
-    expect(container.querySelector('#loading')).toBeNull();
-    expect(container.querySelector('#data')).not.toBeNull();
-    expect(container.querySelector('#data')!.textContent).toBe('Hello World');
-  });
-
-  it('shows error state when promise rejects', async () => {
-    let rejectPromise!: (reason: unknown) => void;
-    const promise = new Promise<string>((_res, rej) => {
-      rejectPromise = rej;
-    });
-
-    function* DataComp() {
-      const data = yield* useResolve(
-        {
-          fn: (_signal) => promise,
-          loading: createElement('span', { id: 'loading' }, 'Loading…'),
-          error: createElement('span', { id: 'error' }, 'Error'),
-        },
-        [],
-      );
-      return createElement('span', { id: 'data' }, data);
-    }
-
-    render(createElement(DataComp as never, {}), container);
-    expect(container.querySelector('#loading')).not.toBeNull();
-
-    rejectPromise(new Error('network error'));
-    await promise.catch(() => {}); // wait for rejection to propagate
-
-    expect(container.querySelector('#error')).not.toBeNull();
-    expect(container.querySelector('#data')).toBeNull();
-  });
-
-  it('useResolve can coexist with useState in the same component', async () => {
-    let resolvePromise!: (data: string) => void;
-    const promise = new Promise<string>((res) => {
-      resolvePromise = res;
-    });
-    let setLabel: ((v: string) => void) | null = null;
-
-    function* DataComp() {
-      const [label, sl] = yield* useState('prefix');
-      setLabel = sl;
-      const data = yield* useResolve(
-        {
-          fn: (_signal) => promise,
-          loading: createElement('span', { id: 'loading' }, 'Loading…'),
-          error: createElement('span', null, 'Error'),
-        },
-        [],
-      );
-      return createElement('p', { id: 'result' }, `${label}:${data}`);
-    }
-
-    render(createElement(DataComp as never, {}), container);
-    expect(container.querySelector('#loading')).not.toBeNull();
-
-    resolvePromise('world');
-    await promise;
-
-    expect(container.querySelector('#result')!.textContent).toBe('prefix:world');
-
-    setLabel!('updated');
-    expect(container.querySelector('#result')!.textContent).toBe('updated:world');
-  });
-
-  it('useState change while promise is pending triggers fresh run and shows correct state after resolve', async () => {
-    let resolvePromise!: (data: string) => void;
-    const promise = new Promise<string>((res) => {
-      resolvePromise = res;
-    });
-    let setLabel: ((v: string) => void) | null = null;
-
-    function* DataComp() {
-      const [label, sl] = yield* useState('prefix');
-      setLabel = sl;
-      const data = yield* useResolve(
-        {
-          fn: (_signal) => promise,
-          loading: createElement('span', { id: 'loading' }, 'Loading…'),
-          error: createElement('span', null, 'Error'),
-        },
-        [],
-      );
-      return createElement('p', { id: 'result' }, `${label}:${data}`);
-    }
-
-    render(createElement(DataComp as never, {}), container);
-    expect(container.querySelector('#loading')).not.toBeNull();
-
-    // Change state WHILE the promise is still pending
-    setLabel!('updated');
-    // Still loading, but label should be reflected after resolve
-    expect(container.querySelector('#loading')).not.toBeNull();
-
-    resolvePromise('world');
-    await promise;
-
-    // The fresh run after setLabel captured 'updated'; resume uses that generator
-    expect(container.querySelector('#result')!.textContent).toBe('updated:world');
-  });
-
-  it('useResolve re-runs when deps change', async () => {
-    let resolveFirst!: (data: string) => void;
-    let resolveSecond!: (data: string) => void;
-    const firstPromise = new Promise<string>((res) => {
-      resolveFirst = res;
-    });
-    const secondPromise = new Promise<string>((res) => {
-      resolveSecond = res;
-    });
-
-    let setId: ((v: number) => void) | null = null;
-    let fetchCount = 0;
-
-    function* DataComp() {
-      const [id, si] = yield* useState(1);
-      setId = si;
-      const data = yield* useResolve(
-        {
-          fn: (_signal) => {
-            fetchCount++;
-            return id === 1 ? firstPromise : secondPromise;
-          },
-          loading: createElement('span', { id: 'loading' }, 'Loading…'),
-          error: createElement('span', null, 'Error'),
-        },
-        [id],
-      );
-      return createElement('p', { id: 'result' }, `${id}:${data}`);
-    }
-
-    render(createElement(DataComp as never, {}), container);
-    expect(fetchCount).toBe(1);
-    expect(container.querySelector('#loading')).not.toBeNull();
-
-    resolveFirst('user1');
-    await firstPromise;
-    expect(container.querySelector('#result')!.textContent).toBe('1:user1');
-
-    // Change the dep – should re-run the promise
-    setId!(2);
-    expect(fetchCount).toBe(2);
-    expect(container.querySelector('#loading')).not.toBeNull();
-
-    resolveSecond('user2');
-    await secondPromise;
-    expect(container.querySelector('#result')!.textContent).toBe('2:user2');
-  });
-
-  it('stale promise result is ignored when deps change before it resolves', async () => {
-    let resolveFirst!: (data: string) => void;
-    let resolveSecond!: (data: string) => void;
-    const firstPromise = new Promise<string>((res) => {
-      resolveFirst = res;
-    });
-    const secondPromise = new Promise<string>((res) => {
-      resolveSecond = res;
-    });
-
-    let setId: ((v: number) => void) | null = null;
-
-    function* DataComp() {
-      const [id, si] = yield* useState(1);
-      setId = si;
-      const data = yield* useResolve(
-        {
-          fn: (_signal) => (id === 1 ? firstPromise : secondPromise),
-          loading: createElement('span', { id: 'loading' }, 'Loading…'),
-          error: createElement('span', null, 'Error'),
-        },
-        [id],
-      );
-      return createElement('p', { id: 'result' }, `${id}:${data}`);
-    }
-
-    render(createElement(DataComp as never, {}), container);
-    expect(container.querySelector('#loading')).not.toBeNull();
-
-    // Change dep before first promise resolves
-    setId!(2);
-    expect(container.querySelector('#loading')).not.toBeNull();
-
-    // Resolve second promise first
-    resolveSecond('user2');
-    await secondPromise;
-    expect(container.querySelector('#result')!.textContent).toBe('2:user2');
-
-    // Now resolve the stale first promise – should NOT update the DOM
-    resolveFirst('user1');
-    await firstPromise;
-    expect(container.querySelector('#result')!.textContent).toBe('2:user2');
-  });
-
-  it('aborts the previous AbortSignal when deps change', async () => {
-    const abortedSignals: AbortSignal[] = [];
-    let setId: ((v: number) => void) | null = null;
-
-    function* DataComp() {
-      const [id, si] = yield* useState(1);
-      setId = si;
-      yield* useResolve(
-        {
-          fn: (signal) => {
-            signal.addEventListener('abort', () => abortedSignals.push(signal));
-            return new Promise(() => {}); // never resolves
-          },
-          loading: createElement('span', { id: 'loading' }, 'Loading…'),
-          error: createElement('span', null, 'Error'),
-        },
-        [id],
-      );
-      return createElement('span', {}, 'done');
-    }
-
-    render(createElement(DataComp as never, {}), container);
-    expect(abortedSignals).toHaveLength(0);
-
-    // Changing deps should abort the first signal and start a new fetch.
-    setId!(2);
-    expect(abortedSignals).toHaveLength(1);
-    expect(abortedSignals[0].aborted).toBe(true);
-  });
-
-  it('aborts the AbortSignal when the component unmounts', async () => {
-    let capturedSignal: AbortSignal | null = null;
-    let setShow: ((v: boolean) => void) | null = null;
-
-    function* Inner() {
-      yield* useResolve(
-        {
-          fn: (signal) => {
-            capturedSignal = signal;
-            return new Promise(() => {}); // never resolves
-          },
-          loading: createElement('span', { id: 'loading' }, 'Loading…'),
-          error: createElement('span', null, 'Error'),
-        },
-        [],
-      );
-      return createElement('span', {}, 'done');
-    }
-
-    function* Outer() {
-      const [show, ss] = yield* useState(true);
-      setShow = ss;
-      return show ? createElement(Inner as never, {}) : null;
-    }
-
-    render(createElement(Outer as never, {}), container);
-    expect(capturedSignal).not.toBeNull();
-    expect(capturedSignal!.aborted).toBe(false);
-
-    // Unmount Inner by hiding it – the AbortSignal should be aborted.
-    setShow!(false);
-    expect(capturedSignal!.aborted).toBe(true);
-  });
-});
-
-describe('render – generator components with useResolveRaw', () => {
-  let container: HTMLElement;
-
-  beforeEach(() => {
-    container = document.createElement('div');
-    document.body.appendChild(container);
-  });
-
-  afterEach(() => {
-    document.body.removeChild(container);
-  });
-
-  it('renders loading state while promise is pending', async () => {
-    let resolvePromise!: (data: string) => void;
-    const promise = new Promise<string>((res) => {
-      resolvePromise = res;
-    });
-
-    function* DataComp() {
-      const p = yield* useMemo(() => promise, []);
-      const { data, loading } = yield* useResolveRaw<string>(p);
-      if (loading) return createElement('span', { id: 'loading' }, 'Loading…');
-      return createElement('span', { id: 'data' }, data);
-    }
-
-    render(createElement(DataComp as never, {}), container);
-    expect(container.querySelector('#loading')).not.toBeNull();
-    expect(container.querySelector('#data')).toBeNull();
-
-    resolvePromise('Hello');
-    await promise;
-
-    expect(container.querySelector('#loading')).toBeNull();
-    expect(container.querySelector('#data')).not.toBeNull();
-    expect(container.querySelector('#data')!.textContent).toBe('Hello');
-  });
-
-  it('renders error state when promise rejects', async () => {
-    let rejectPromise!: (reason: unknown) => void;
-    const promise = new Promise<string>((_res, rej) => {
-      rejectPromise = rej;
-    });
-
-    function* DataComp() {
-      const p = yield* useMemo(() => promise, []);
-      const { data, loading, error } = yield* useResolveRaw<string, Error>(p);
-      if (loading) return createElement('span', { id: 'loading' }, 'Loading…');
-      if (error) return createElement('span', { id: 'error' }, error.message);
-      return createElement('span', { id: 'data' }, data);
-    }
-
-    render(createElement(DataComp as never, {}), container);
-    expect(container.querySelector('#loading')).not.toBeNull();
-
-    rejectPromise(new Error('network error'));
-    await promise.catch(() => {});
-
-    expect(container.querySelector('#error')).not.toBeNull();
-    expect(container.querySelector('#error')!.textContent).toBe('network error');
-    expect(container.querySelector('#data')).toBeNull();
-  });
-
-  it('re-fetches when the promise reference changes', async () => {
-    let resolveFirst!: (data: string) => void;
-    let resolveSecond!: (data: string) => void;
-    const firstPromise = new Promise<string>((res) => {
-      resolveFirst = res;
-    });
-    const secondPromise = new Promise<string>((res) => {
-      resolveSecond = res;
-    });
-    let setId: ((v: number) => void) | null = null;
-
-    function* DataComp() {
-      const [id, si] = yield* useState(1);
-      setId = si;
-      const p = yield* useMemo(() => (id === 1 ? firstPromise : secondPromise), [id]);
-      const { data, loading } = yield* useResolveRaw<string>(p);
-      if (loading) return createElement('span', { id: 'loading' }, 'Loading…');
-      return createElement('p', { id: 'result' }, `${id}:${data}`);
-    }
-
-    render(createElement(DataComp as never, {}), container);
-    expect(container.querySelector('#loading')).not.toBeNull();
-
-    resolveFirst('user1');
-    await firstPromise;
-    expect(container.querySelector('#result')!.textContent).toBe('1:user1');
-
-    setId!(2);
-    expect(container.querySelector('#loading')).not.toBeNull();
-
-    resolveSecond('user2');
-    await secondPromise;
-    expect(container.querySelector('#result')!.textContent).toBe('2:user2');
-  });
-
-  it('ignores stale promise result when promise reference changes', async () => {
-    let resolveFirst!: (data: string) => void;
-    let resolveSecond!: (data: string) => void;
-    const firstPromise = new Promise<string>((res) => {
-      resolveFirst = res;
-    });
-    const secondPromise = new Promise<string>((res) => {
-      resolveSecond = res;
-    });
-    let setId: ((v: number) => void) | null = null;
-
-    function* DataComp() {
-      const [id, si] = yield* useState(1);
-      setId = si;
-      const p = yield* useMemo(() => (id === 1 ? firstPromise : secondPromise), [id]);
-      const { data, loading } = yield* useResolveRaw<string>(p);
-      if (loading) return createElement('span', { id: 'loading' }, 'Loading…');
-      return createElement('p', { id: 'result' }, `${id}:${data}`);
-    }
-
-    render(createElement(DataComp as never, {}), container);
-
-    setId!(2);
-
-    resolveSecond('user2');
-    await secondPromise;
-    expect(container.querySelector('#result')!.textContent).toBe('2:user2');
-
-    resolveFirst('user1');
-    await firstPromise;
-    // Stale result must not overwrite the current render
-    expect(container.querySelector('#result')!.textContent).toBe('2:user2');
-  });
-});
-
-describe('buildNode', () => {
-  it('returns a text node for strings', () => {
-    const node = buildNode('hello');
-    expect(node).toBeInstanceOf(Text);
-    expect(node.textContent).toBe('hello');
-  });
-
-  it('returns a text node for numbers', () => {
-    const node = buildNode(42);
-    expect(node).toBeInstanceOf(Text);
-    expect(node.textContent).toBe('42');
-  });
-
-  it('returns an empty text node for null', () => {
-    const node = buildNode(null);
-    expect(node).toBeInstanceOf(Text);
-    expect(node.textContent).toBe('');
-  });
-
-  it('returns an empty text node for false', () => {
-    const node = buildNode(false);
-    expect(node).toBeInstanceOf(Text);
-    expect(node.textContent).toBe('');
-  });
-});
-
-describe('shown prop', () => {
-  let container: HTMLElement;
-
-  beforeEach(() => {
-    container = document.createElement('div');
-    document.body.appendChild(container);
-  });
-
-  afterEach(() => {
-    document.body.removeChild(container);
-  });
-
-  it('renders an HTML element when shown is true', () => {
-    render(createElement('div', { $shown: true }, 'visible'), container);
-    expect(container.querySelector('div')).not.toBeNull();
-    expect(container.querySelector('div')!.textContent).toBe('visible');
-  });
-
-  it('does not render an HTML element when shown is false', () => {
-    render(createElement('div', { $shown: false }, 'hidden'), container);
-    expect(container.querySelector('div')).toBeNull();
-  });
-
-  it('renders an HTML element when shown is omitted (defaults to shown)', () => {
-    render(createElement('div', null, 'visible'), container);
-    expect(container.querySelector('div')).not.toBeNull();
-  });
-
-  it('does not set shown as a DOM attribute', () => {
-    render(createElement('div', { $shown: true }, 'visible'), container);
-    const el = container.querySelector('div')!;
-    expect(el.hasAttribute('$shown')).toBe(false);
-  });
-
-  it('renders a plain function component when shown is true', () => {
-    function Greeting() {
-      return createElement('p', null, 'hello');
-    }
-    render(
-      createElement('div', null, createElement(Greeting as never, { $shown: true })),
-      container,
-    );
-    expect(container.querySelector('p')).not.toBeNull();
-  });
-
-  it('does not render a plain function component when shown is false', () => {
-    function Greeting() {
-      return createElement('p', null, 'hello');
-    }
-    render(
-      createElement('div', null, createElement(Greeting as never, { $shown: false })),
-      container,
-    );
-    expect(container.querySelector('p')).toBeNull();
-  });
-
-  it('renders a generator component when shown is true', () => {
-    function* Counter() {
-      return createElement('p', null, 'counter');
-    }
-    render(
-      createElement('div', null, createElement(Counter as never, { $shown: true })),
-      container,
-    );
-    expect(container.querySelector('p')).not.toBeNull();
-  });
-
-  it('does not render a generator component when shown is false', () => {
-    function* Counter() {
-      return createElement('p', null, 'counter');
-    }
-    render(
-      createElement('div', null, createElement(Counter as never, { $shown: false })),
-      container,
-    );
-    expect(container.querySelector('p')).toBeNull();
-  });
-
-  it('unmounts an HTML element when shown changes from true to false', () => {
-    let setShown: ((v: boolean) => void) | null = null;
-
-    function* Wrapper() {
-      const [$shown, setS] = yield* useState(true);
-      setShown = setS;
-      return createElement('div', { $shown }, 'content');
-    }
-
-    render(createElement(Wrapper as never, {}), container);
-    expect(container.querySelector('div')).not.toBeNull();
-
-    setShown!(false);
-    expect(container.querySelector('div')).toBeNull();
-  });
-
-  it('mounts an HTML element when shown changes from false to true', () => {
-    let setShown: ((v: boolean) => void) | null = null;
-
-    function* Wrapper() {
-      const [$shown, setS] = yield* useState(false);
-      setShown = setS;
-      return createElement('div', { $shown }, 'content');
-    }
-
-    render(createElement(Wrapper as never, {}), container);
-    expect(container.querySelector('div')).toBeNull();
-
-    setShown!(true);
-    expect(container.querySelector('div')).not.toBeNull();
-    expect(container.querySelector('div')!.textContent).toBe('content');
-  });
-
-  it('unmounts a component when shown changes from true to false', () => {
-    let setShown: ((v: boolean) => void) | null = null;
-
-    function* Inner() {
-      return createElement('p', null, 'inner');
-    }
-
-    function* Wrapper() {
-      const [$shown, setS] = yield* useState(true);
-      setShown = setS;
-      return createElement(Inner as never, { $shown });
-    }
-
-    render(createElement(Wrapper as never, {}), container);
-    expect(container.querySelector('p')).not.toBeNull();
-
-    setShown!(false);
-    expect(container.querySelector('p')).toBeNull();
-  });
-
-  it('mounts a component when shown changes from false to true', () => {
-    let setShown: ((v: boolean) => void) | null = null;
-
-    function* Inner() {
-      return createElement('p', null, 'inner');
-    }
-
-    function* Wrapper() {
-      const [$shown, setS] = yield* useState(false);
-      setShown = setS;
-      return createElement(Inner as never, { $shown });
-    }
-
-    render(createElement(Wrapper as never, {}), container);
-    expect(container.querySelector('p')).toBeNull();
-
-    setShown!(true);
-    expect(container.querySelector('p')).not.toBeNull();
-    expect(container.querySelector('p')!.textContent).toBe('inner');
-  });
-});
-
-describe('render – useEffect', () => {
-  let container: HTMLElement;
-
-  beforeEach(() => {
-    container = document.createElement('div');
-    document.body.appendChild(container);
-  });
-
-  afterEach(() => {
-    document.body.removeChild(container);
-  });
-
-  it('runs effect after initial mount', () => {
-    const calls: string[] = [];
-
-    function* Comp() {
-      yield* useEffect(() => {
-        calls.push('effect');
-      }, []);
-      return createElement('span', {}, 'hi');
-    }
-
-    render(createElement(Comp as never, {}), container);
-    expect(calls).toEqual(['effect']);
-  });
-
-  it('does not re-run effect when deps are unchanged', () => {
-    const calls: string[] = [];
-    let setCount: ((v: number) => void) | null = null;
-
-    function* Comp() {
-      const [count, sc] = yield* useState(0);
-      setCount = sc;
-      yield* useEffect(() => {
-        calls.push('effect');
-      }, []); // empty deps — should only run once
-      return createElement('span', {}, String(count));
-    }
-
-    render(createElement(Comp as never, {}), container);
-    expect(calls).toEqual(['effect']);
-
-    setCount!(1);
-    setCount!(2);
-    expect(calls).toEqual(['effect']); // still only once
-  });
-
-  it('re-runs effect and calls previous cleanup when deps change', () => {
-    const log: string[] = [];
-    let setId: ((v: number) => void) | null = null;
-
-    function* Comp() {
-      const [id, si] = yield* useState(1);
-      setId = si;
-      yield* useEffect(() => {
-        log.push(`effect:${id}`);
-        return () => log.push(`cleanup:${id}`);
-      }, [id]);
-      return createElement('span', {}, String(id));
-    }
-
-    render(createElement(Comp as never, {}), container);
-    expect(log).toEqual(['effect:1']);
-
-    setId!(2);
-    expect(log).toEqual(['effect:1', 'cleanup:1', 'effect:2']);
-
-    setId!(3);
-    expect(log).toEqual(['effect:1', 'cleanup:1', 'effect:2', 'cleanup:2', 'effect:3']);
-  });
-
-  it('calls cleanup on unmount', () => {
-    const log: string[] = [];
-    let setShow: ((v: boolean) => void) | null = null;
-
-    function* Inner() {
-      yield* useEffect(() => {
-        log.push('mount');
-        return () => log.push('unmount');
-      }, []);
-      return createElement('span', {}, 'inner');
-    }
-
-    function* Outer() {
-      const [show, ss] = yield* useState(true);
-      setShow = ss;
-      return show ? createElement(Inner as never, {}) : null;
-    }
-
-    render(createElement(Outer as never, {}), container);
-    expect(log).toEqual(['mount']);
-
-    setShow!(false);
-    expect(log).toEqual(['mount', 'unmount']);
-  });
-
-  it('does not run effect while generator is paused in useRender', () => {
-    const log: string[] = [];
-    let capturedResume: ((v: string) => void) | null = null;
-
-    function* Comp() {
-      yield* useEffect(() => {
-        log.push('effect');
-      }, []);
-      const answer = yield* (function* (): Generator<unknown, string, unknown> {
-        // Inline useRender-like pause: yield a VNode to pause the generator
-        const caps = (yield {
-          type: Symbol.for('yielderact.useRender.test'),
-        }) as null;
-        return caps as unknown as string;
-      })();
-      return createElement('span', {}, answer);
-    }
-
-    // We cannot easily test useRender interaction without full plumbing,
-    // so instead test via the abort-signal / pending effect flow with a
-    // simpler approach: effect runs only when generator finishes.
-    // Just verify effect ran after a full render cycle.
-    function* Simple() {
-      yield* useEffect(() => {
-        log.push('ran');
-      }, []);
-      return createElement('span', {}, 'ok');
-    }
-
-    render(createElement(Simple as never, {}), container);
-    expect(log).toEqual(['ran']);
-    void Comp; // silence unused warning
-  });
-});
-
-describe('render – HTML defaults', () => {
-  it('sets button type to "button" when not specified', () => {
-    const node = buildNode(createElement('button', {}, 'Click')) as HTMLButtonElement;
-    expect(node.getAttribute('type')).toBe('button');
-  });
-
-  it('preserves explicit button type', () => {
-    const node = buildNode(
-      createElement('button', { type: 'submit' }, 'Submit'),
-    ) as HTMLButtonElement;
-    expect(node.getAttribute('type')).toBe('submit');
-  });
-
-  it('warns when <a target="_blank"> has no rel', () => {
-    const warn = jest.spyOn(console, 'warn').mockImplementation(() => {});
-    buildNode(createElement('a', { href: 'https://example.com', target: '_blank' }, 'link'));
-    expect(warn).toHaveBeenCalledWith(expect.stringContaining('noopener'));
-    warn.mockRestore();
-  });
-
-  it('does not warn when <a target="_blank"> has any rel value', () => {
-    const warn = jest.spyOn(console, 'warn').mockImplementation(() => {});
-    buildNode(
-      createElement(
-        'a',
-        { href: 'https://example.com', target: '_blank', rel: 'noopener noreferrer' },
-        'link',
-      ),
-    );
-    buildNode(
-      createElement(
-        'a',
-        { href: 'https://example.com', target: '_blank', rel: 'noreferrer' },
-        'link',
-      ),
-    );
-    expect(warn).not.toHaveBeenCalled();
-    warn.mockRestore();
-  });
-
-  it('does not warn for <a> without target="_blank"', () => {
-    const warn = jest.spyOn(console, 'warn').mockImplementation(() => {});
-    buildNode(createElement('a', { href: 'https://example.com' }, 'link'));
-    expect(warn).not.toHaveBeenCalled();
-    warn.mockRestore();
-  });
-});
